STORM-1931 : Share mapper and selector in Storm-Kafka
