+++ external/storm-kafka/src/jvm/org/apache/storm/kafka/selector/FieldIndexTopicSelector.java
@@ -15,9 +15,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.storm.kafka.bolt.selector;
+package org.apache.storm.kafka.selector;
 
-import org.apache.storm.tuple.Tuple;
+import org.apache.storm.tuple.ITuple;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -37,7 +37,7 @@ public class FieldIndexTopicSelector implements KafkaTopicSelector {
     }
 
     @Override
-    public String getTopic(Tuple tuple) {
+    public String getTopic(ITuple tuple) {
         if (fieldIndex < tuple.size()) {
             return tuple.getString(fieldIndex);
         } else {
