+++ external/storm-kafka/src/test/org/apache/storm/kafka/TridentKafkaTopology.java
@@ -20,13 +20,12 @@ package org.apache.storm.kafka;
 import org.apache.storm.Config;
 import org.apache.storm.LocalCluster;
 import org.apache.storm.generated.StormTopology;
+import org.apache.storm.kafka.mapper.FieldNameBasedTupleToKafkaMapper;
+import org.apache.storm.kafka.selector.DefaultTopicSelector;
 import org.apache.storm.tuple.Fields;
 import org.apache.storm.tuple.Values;
-import com.google.common.collect.ImmutableMap;
 import org.apache.storm.kafka.trident.TridentKafkaStateFactory;
 import org.apache.storm.kafka.trident.TridentKafkaUpdater;
-import org.apache.storm.kafka.trident.mapper.FieldNameBasedTupleToKafkaMapper;
-import org.apache.storm.kafka.trident.selector.DefaultTopicSelector;
 import org.apache.storm.trident.Stream;
 import org.apache.storm.trident.TridentTopology;
 import org.apache.storm.trident.testing.FixedBatchSpout;
